Code/App/backend/requirements.txt
Django==3.0.8
djangorestframework==3.11.0
joblib==0.15.1
numpy==1.18.4
PyStemmer==2.0.1
scikit_learn==0.23.1
wikipedia==1.4.0
django-cors-headers==3.4.0
